fix external-schema path in cached model generation (#3120)
* Update SchemaLocator.php

Replaces the use of a relative path with an absolute path. For example, when the generation of models in the cache is called from the Thelia php console, it uses the path local/config/shema.xml. Whereas if the generation of models in the cache is called from an http request, it uses the path web/local/config/shema.xml.

* Update SchemaLocator.php

fix code style